[create]comments and ordering by likes

+++ post/views.py
@@ -1,12 +1,14 @@
 from django.shortcuts import render
 
 from rest_framework.response import Response
-from .models import Post
+from .models import Post, Like
 from rest_framework.decorators import api_view
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework import status
 from rest_framework.views import APIView
 from .serializers import PostSerializer
+from tag.models import Tag
+from django.db.models import Count
 
 # Create your views here.
 
@@ -31,16 +33,29 @@ class PostListView(APIView):
 
 		### 얘네가 class inner function 들! ###
     def get(self, request): 
-        posts = Post.objects.all()
+        # posts = Post.objects.all()
+        posts = Post.objects.annotate(like_count=Count('like_users')).order_by('-like_count')
         serializer = PostSerializer(posts, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
+        author = request.user
         title = request.data.get('title')
         content = request.data.get('content')
+        tag_ids = request.data.get('tags')
+        
+        if not author.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+        
         if not title or not content:
             return Response({"detail": "[title, content] fields missing."}, status=status.HTTP_400_BAD_REQUEST)
-        post = Post.objects.create(title=title, content=content)#cf)create = make + save(no need to save)
+        
+        for tag_id in tag_ids:
+            if not Tag.objects.filter(id=tag_id).exists():
+                return Response({"detail": "Provided tag not found."}, status=status.HTTP_404_NOT_FOUND)
+        
+        post = Post.objects.create(title=title, content=content, author=author)#cf)create = make + save(no need to save)
+        post.tags.set(tag_ids)
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
@@ -75,5 +90,48 @@ class PostDetailView(APIView):
             post = Post.objects.get(id=post_id)
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+        
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+        
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+    
+
+    def patch(self, request, post_id):
+            try:
+                post = Post.objects.get(id=post_id)
+            except:
+                return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+            
+            if request.user != post.author:
+                return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
+
+            serializer = PostSerializer(post, data=request.data, partial=True)
+            if not serializer.is_valid():
+                return Response({"detail": "data validation error"}, status=status.HTTP_400_BAD_REQUEST)
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+    
+class LikeView(APIView):
+    def post(self, request, post_id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        ### 1 ###
+        try:
+            post = Post.objects.get(id=post_id)
+        except:
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        ### 2 ###
+        like_list = post.like_set.filter(user=request.user)
+
+        ### 3 ###
+        if like_list.count() > 0:
+            post.like_set.get(user=request.user).delete()
+        else:
+            Like.objects.create(user=request.user, post=post)
+
+        serializer = PostSerializer(instance=post)
+        return Response(serializer.data, status=status.HTTP_200_OK)
